Initial checkin for the Steam monitor and its connection filter DLL
The call into the filter DLL can be reasonably general, calling any entry
point and passing a (currently string only) parameter to it. The receiver
now uses hand-linking to parse the passed-in address specification to form
the address to direct Steam towards, and now the unloading is done through
an explicit side call rather than a second entry point in the shim.

This means that the shim can now always be unloaded, and the unload behaviour
will be simpler and generally more reliable; it'll be possible to leave the
filter DLL attached when the monitor exits, and then later running the
monitor again will have no ill effects (modulo any intermediate version
upgrade, which is a scenario that's tricky - that will best be handled by
the installer via uninstall-reinstall).

Add version information, an icon, notification area support, auto-launch
configurable from the context menu, a BSD license, an NSIS-based installer,
configurable server IP (in the registry for now).