Update to 0.2.1 with various cosmetic fixes.
Properly work around the TrackPopupMenu bug with not dismissing by ensuring
the main window has been made visible at least once, otherwise the auto-
close for the pop-up menu does not work.

The tooltip for the application is now a resource, and there's a simple
modeless "about" dialog shown for grins since it's handy to make version
numbers easy to see.