Resolve some tedious problems with Win7 UAC for 0.3.2
This mainly deals with some awkward circumlocutions required to work with UAC
the way I want the app to work, where it should not be elevated.

This means arranging for the installer to launch an instance of the app non-
elevated after it has installed (instead of elevated as before) and to use
the NSIS UAC plugin to relaunch an elevated instance of the installer to do
the necessary work of copying the program files and manipulating the keys
under HKLM to register with the Windows uninstall system.

This necessitates moving the program settings from HKEY_LOCAL_MACHINE to the
HKEY_CURRENT_USER hive so that the settings can be changed by the non-elevated
instance of the application running in the system notification area.