Add some experimental HTTP filtering, make unloading more robust.
The alternative HTTP content-downloading system new to Steam (which in theory
is supposed to go with some kind of delta-patch support) is problematic for
filtering purposes. This adds support for filtering out that traffic too.

In normal Steam downloads, the server IP's to use for a post 27030 download
are sent from an internal Steam server on port 27033. For some products, the
download it instead done over HTTP in several stages, using a different CDN
entirely! For such products, there is not (and very likely cannot be, until
Steam makes some substantial changes) ISP-level support for private replicas.

Since it's not possible for customers to tell which are HTTP-only titles and
which are delivered over the classic Steam CDN, this means a nasty surprise.

So, I've added support for blocking the newer HTTP CDN. This works due to the
fact that unlike the classic CDN's port 27033 service to get a server list as
a raw IP+port list in binary format, the newer HTTP CDN process starts with a
JSON result document which contains the hostnames of the servers in the HTTP
CDN.

Since the steam client then resolves those names as part of the connection
process (so presumably Valve are planning to add their future load-balancing
via DNS results, as a lot of cloud-hosted Web services tend to do these days)
although we don't filter the connections we can filter the DNS queries.

At present, the servers on the separate Steam HTTP CDN have distinct names
and so we can block resolution of those names to prevent HTTP downloads. This
is not ideal, but since it seems that the choice of HTTP or the classic Steam
CDN is a global product-level switch, there's nothing much that can be done
other than this.

Adding this block shows a difference with HTTP-level filtering; once we block
the resolution of the HTTP CDN entirely, the download initiation stops at the
"preparing files" stage before adding the game to the download list (unlike
downloads via the traditional CDN). This means that there's a certain level
of visibility for titles which are on HTTP, although presumably I should add
some kind of pop-up notification that the active HTTP blocking has triggered
and users will have to turn off filtering to download things.

This checkin also includes a small fix to the top-level window monitoring as
this could get borked a little by Steam client updates (oops) leaving the
filter DLL stuck as in-use until Steam is quit, and more important the
filtering not being done against the new Steam instance.