Add a small Google App Engine support service.
The main point of the GAE service is to provide a stable URL pointing at the
most recent version of the Steam Limiter; it can return the version string,
redirect to the download page, or redirect to the download itself.

As well, there is also some simple code to use ISP identification to work out
which of the supported ISPs a request is coming from. The intention here is
to be able to add an installation-time helper to fetch the correct default
Steam server to filter based on the installing user's ISP; to do this I get
a subset of the GeoIP data available at http://www.maxmind.com/app/asnum and
use some scripts to bake this data into a Python script so that it's easy to
look up in the App Engine script (and won't incur any Datastore access).

http://steam-limiter.appspot.com contains the running instance of the server
code; having created it and tested it, the next stage is to look at building
various WSH scripts to probe the data from the client side, most likely with
MSXML2.XmlHttpRequest (version 6 or 3, per the MS ProgID guidelines). The
reason for using WSH script for this is mainly that it's the lightest way to
write this in terms of both download size and load in the constantly-running
parts like the tray applet.