Add update checking to the monitor app and auto-ISP selection on install.
The previous checkin added a simple Google App Engine service which returns
information about the current most recent version of the application and has
a GeoIP lookup facility to determine which ISP is being used and thus what
the correct Steam server to filter to for the user is.

This checkin has the installer script calling out to a script which does an
XmlHttpRequest to the GAE webservice on steam-limiter.appspot.com and uses
it to select the initial server to filter.

It also expands the monitor applet with an "upgraded version available"
alternative to the "about" dialog, along with an "upgrade now" button which
will invoke the webservice to get the path to the latest installer, download
the installer, and then runs it.

The main intention here is to stop individual users from having to check the
Google Code project page. It may be that it's also a good idea to have the
"upgrade now" option appear in the context menu and perhaps to decorate the
context-menu icon with an overlay or generate a balloon (the last being the
approach recommended for Win7 where overlay icons aren't very visible, but
probably needing a 30-day timer much like the upgrade network check).