Add a flexible rule parser allowing multiple target IPs to rewrite to.
This is not yet the full v0.5.0 release but it's most of the planned work
less some UI adjustments while I decide the ideal approach to deciding on
how to upgrade rules from the older form to the current one and/or to allow
for things like re-running the detection rules and easy uploading of custom
rules as a way to support user-driven improvements to the rulebase.

Since there are already folks in Australia using the limiter, this therefore
is the first set of changes to support them properly. In Australia (where
peering arrangements are more robust than in New Zealand) it is common for
ISPs to either support each other's Steam servers and for some of the
unmetered servers to be unfiltered, which is different from New Zealand.

In this circumstance, it's best for the limiter to be able to select the IP
addresses to rewrite to from a list rather than a single one, so that the
Steam client can still do parallel downloading. Also, the DNS rewrite rules
can now return an alternative IP instead of just a "not found", so this can
potentially be used to fake out the HTTP download system if the unmetered
server has port 80 open (although not many of them do, particularly the
filtered ones).

Although just supporting multiple IPs is relatively simple, this change is a
bit bigger and potentially allows the system to do more extensive kinds of
rewrite actions on different ports. This is partly in anticipation of how
Valve are likely to extend their HTTP CDN to support ISP-level servers, and
also just as an obvious generalization of the filtering system to make it
easy to use for applications other than Steam (since the general approach is
potentially very useful in things like QA automation).

Between being able to use rule specifications to rewrite DNS query results
and to rewrite the address and port in connect calls (and other Windows
Sockets APIs, in the more general case) to do more substantial actions such
as to trigger scripted process launches for logging or to proxy just one of
an application's connections.

This is useful because in continuous build and test automation, it is hard
to exercise network functions of programs in live code, without needing some
network virtualization or separate test paths, and fault injection for
exercising recovery code for network failures is particularly tedious. The
kind of approaches used here make this rather simpler; in particular, by
virtue of doing this using API hooking means that it's easy to run such
basic network test automation on developer builds.

While the Steam Limiter won't need to use all that just yet, having the code
framework present to make it easy to add is useful, and it's not actually
that much more complex to do it this way, since it's really a generalization
and unification of the DNS filtering with the connect filtering.

The fancy new rule parser is happy to treat the old-style simple IPv4 spec
as a replacement for a global pattern, so rules from existing installs will
work just fine. Basically the only incompatible change is for the web service
as the rules for Australian ISPs will be more complex than the prev-v0.5
limiter can use.

I therefore need to include some form of versioning in the setfilter.js user
agent string, and to add a parameter to the filter URL which indicates the
limiter client can support the service returning multiple-IP default
rules.

Currently this is under test, but I'll be adding more before the actual
v0.5.0 release, mostly including some changes to the setfilter.js script
to add version data to the user-agent string and perhaps some more depending
on how I decide to handle some of the upgrade issues.

In addition, there are a number of UI changes and storage changes which are
going to be made for the final v0.5 release based on the fact that the rules
for filtering are becoming more complex and need to reside in the webservice
for easier updating without requiring new code to be deployed (also see
issue #8 for having separate profiles).