Expand the in-progress v0.5 code with bugfixes and creature comforts.
This includes a resolution for item #9 (suspending the limiter) and a proper
hyperlink control for the about box intended to support various other links
to online help documents in the new dialogs I will be making for v0.5,
which also simplifies the context menu. The config dialogs now appear in the
taskbar but still have no caption, but support full-window dragging.

The injection system also now includes new fail-safe code to help deal with
orphaned processes leaving injected DLLs around (which doesn't happen at all
in normal use but it really annoying when developing, and a good safety move
in general) by giving them extra encouragement to unload.

This also fills in a few cases in the new filter syntax so that blocks rules
and success rules can exist anywhere and override later ones, so that there
is (as intended) now a way to write rules which mean success and can override
failure patterns like the basic HTTP block as well as force them to specific
values.

One other detail is that when checking for updates the setrule.js script
adds the steam-limiter version number to its user-agent string, which should
help me figure out some details with webservice version upgrades in future.

Also added a custom 404 handler and a robots.txt to the webservice, since
the appengine logs showed a visit from the GoogleBot indexing the service
URLs and it doesn't need to do that.

At this point the main thing needed to release v0.5 is to add the new
profile system; now that the webservice is bedded in and the rule system is
more flexible, it's best to make the auto-detected rules the default and to
even auto-update the rules from the webservice during the new-version update
poll cycle.

It'll still be possible to have custom rules (and in v0.5.1 I'll probably
add the rule-upload support) but the auto-detected profiles will always be
available and this means I can remove the IP list from the executable and
maintain it online in the webservice without having to make any code changes.