Replace the original server selection with a profile management dialog.
This still allows the same manual configuration as in the previous editions,
but to customize the filters you need to select the "Custom" profile, while
the auto-detected profile (which for almost everyone should be all they
need) stays in place.

As well there is an alternate "away" profile which can store a separate
location from the "home" location, and a "disable all traffic" option for
temporarily offlining Steam.

The new profile dialogs also support redetecting the current ISP and making
custom rules which can be uploaded to suggest improvements to the limiter's
rule base.