More suggested mappings
Based on the attachment Angus Wolfcastle (thanks, man) wrote at
http://www.anguswolfcastle.co.cc/steam-1/filtering-methods

Kudos to him.

It seems that since he lists 3FL as (Adam, iiNet, Netspace, Westnet) there's
some additional netblocks for those I need to dig out and include in a
revision, which I'll look at tomorrow.

"Netspace Online" seems to be the right name for the ASN registration for
them from a quick look at the raw database, but I'll double-check with APNIC
tomorrow.

Westnet is probably "Westnet Internet Services".

Adam is the tricky one - LOTS of registrations for variations on that in
names but I'm guessing that "Adam Internet Pty Ltd" is the right one.