Add the University of Waikato as its own ISP.
I am informed that it is peered with Snap! for unmetered Steam for some
student services, but Waikato is really its own ISP (and has a big part in
NZ's internet history with an NSF-related connection from Honolulu providing
early connectivity here).

Student internet is supposedly done through Lightwire, but that's not got
its own netblock as far as I can tell, so is probably just a spinoff that
is using a subnet of the main university netblock.