Thanks to a bug report from a webafrica user, fixed up the HTTP redirection.
This should resolve the crash bug reported as issue 15.

It turns out I had the result data from gethostbyname () in the wrong format
and when testing this against webafrica's content server this provoked a
steam client crash. Fixing this means HTTP downloads against steam.wa.co.za
now work (in my testing, as webafrica's content server is not filtered),
which means that HTTP support should be the default rule for webafrica as
well.