Add InSPire as an ISP, although they have no published IP I can use yet.
I've had an install from InSPire but they run a filtered server, and are so
small they don't even have a forum or any useful FAQ pages which actually
say what it might be.

The other notable item here is a technique for blocking the "CS" type
servers from being involved in HTTP downloads. I've been working with a
WebAfrica customer on this - WA's steam server is currently open to me but
404'ing to WA customers over HTTP, but the testing showed that Valve was
also sending back a long list of "CS" type servers to folks in ZA.

I tried a number of things to filter those but a key problem is that Valve
don't appear to use any of the standard APIs for parsing the URLs sent back
for "CS" type servers, so I can't trap them to filter them.

What I've put in place here is wrapping WSASend (), which is used by the
Steam client's built-in HTTP client (and not the embedded copy of Chrome)
and extracting the request URL from the write - easy, as the HTTP request is
sent in a single API call to Windows Sockets.

The "CS" type servers authenticate with your Steam login cookie, and so this
provides a distinct URL that I can use to filter "CS" servers, since Valve's
GeoIP for selecting them is not just opaque, but clearly messed up - its
suggestions are spectacularly inappropriate for anyplace where there are
unmetered servers to choose instead.

Currently this is testing to see if it works - it's not clear whether I have
to actually fake a response as well, or just can fail the send call. If this
proves to work at blocking CS-type servers then I'll add URL filters to the
filter syntax.