Add the hopefully-final touches to the CS server filtering for v0.5.2
The test of CS server filtering using a hardcoded test worked well, so this
moves the URL filtering into the filter specification strings, along with an
extension to allow URLs to be replaced as well as blocked.

This isn't particularly useful in the context of Steam (and really, could
make good use of a regex-based matching language with backreferences for the
replacement patterns), so it's here more for completeness than practicality.

However, having this in the filter system explicitly is important, since I
think CS server use should be blocked by default for steam-limiter users
and so there needs to be a way of using custom filters to re-enable it.

My expectation is that I'll post this build via a non-Google-Code download
to https://groups.google.com/forum/#!forum/steam-limiter so that it can get
some testing before being transformed into v0.5.2