Fix a couple small bugs in the webservice assistant script.
When there's no connectivity to the webservice (e.g. on a laptop with Wi-Fi
disabled) then the script was trying to use a blank string, which yielded an
error from eval () under WScript; change that to null instead.

Also, when downloading an upgrade package for a newer version on Windows 7,
the upgrade would fail because it was being placed in the current directory
(which will be the "Program Files" directory) - on Win7 this is non-writable,
so instead the upgrade package has to go in the temporary file directory
which is part of the user profile.

In addition, I have adopted a small tweak to the Optus rules for better
performance - Optus evidently don't offer any unmetered services so the
rules exist to constrain it to use more performant servers, although that
would normally be something I'd have expected Steam's default behaviour to
do adequately.