Try and force the limiter dialogs to show on the primary monitor
I've had a user report that on a multi-monitor system sometimes the profile
dialog shows on the secondary monitor and the mouse selection is a little
bit strange with it.

Unfortunately I haven't been able to replicate this (at least using Win7,
since that's the only machine I can set up with multiple monitors), so I'm
left without a great deal to go on. In principle, the API calls I was making
should always open the dialogs centered on the primary monitor, but with XP
it's possible some third-party add-on might be confusing the issue.

Since I'm blind as to the real trigger, what I'll do for now is try and
force the issue, making the dialog positioning explicit and bound to the
primary display by using the monitor information API.

The underlying bug is filed as Issue #17 and once v0.5.2.2 survives a week
or so then I'll include this when I promote it to v0.5.3.0