Updated IPv4 mappings from the most recent ASN database.
Not a lot really has changed, but the underlying data source has some quirks
which are obvious in the latest edition - formerly contiguous ranges have
been split up, and the like. This seems to be best handled by merging ranges
that abut and which map to the same target, and so now the mapping is a bit
smaller overall.

Although the underlying CSV data is refreshed monthly, it seems unlikely to
be important to refresh it more than every month or two.