Republish the v0.5.2.2 test version as v0.5.3.0 to push live shortly.
There is also a small delay added to the installer script to help with fast
machines due to an annoying race condition in Windows where a process can
exit but its executable file won't be free to modify for a short time later.

Also, re-jig the project settings a small amount so that the build project
takes care of prerequisites and timestamps properly, so editing the .NSI
file in VS2010 causes the installer rebuild properly.