Refresh the ASN mapping database with data for March.
As well, add entries for a few ISPs where I've had installs but don't have
any unmetered Steam servers (so it's not clear to me what rules are being
used), and an additional Comcast netblock there has been an install on though
I have yet to hear any rule suggestions appropriate to US users.

In other news, I've had one report from a WebAfrica user who had dual ISPs,
which is something that I hadn't expected to see (in New Zealand such things
aren't permitted with residential DSL, and it's hard to imagine any market
conditions where it'd make economic sense, but it's hard to argue with the
fact someone has done it).

This changelist doesn't include it, but I am contemplating whether there are
any expansions to the detection process that are warranted; in this case I'd
suspect the circumstances mean that only some additional client-side test is
warranted to identify WebAfrica in the case the detection request is coming
from another South African ISP (IS or one of its resellers, or SAIX).