Add a new utility to try and probe whether a port is open on a host.
This exists for the benefit of users in South Africa, where evidently it's
relatively commonplace to have dual ISP connections (where that's largely
unknown in AU and NZ). This means that the existing GAE service to detect
the user's ISP may be correct, but if a user has a static route for an ISP
like WebAfrica then it will detect their "other" ISP as the primary.

WebAfrica don't have any customer-only HTTP services that I can probe for in
the existing script, but they *do* have a customer-only NTTP server. There's
no way to test that using scriptable components (the Collaboration Data
Object system has an NNTP client that can post to, but not read from, NNTP)
so I'm adding this small app to start being able to return additional probe
sequences from the GAE service to the setfilter.js script.

That can then in turn call out to this app; one specific thing that the app
copes with is the existence of the Avast! antivirus program; it contains a
set of transparent proxies for HTTP, SMTP and NNTP traffic, and the proxy is
good at interfering with connections to NNTP servers - they always succeed
but then fail later, after about 10 seconds while the proxy tries the target
on the TLS and then plain NNTP ports. So if Avast! is present, the probe app
has to actually try reading from the connected socket to see whether there's
a 400 response sent.

Currently I'm testing how useful this is, with an eye to then working on the
script and webservice side to integrate this into the regular process in a
5.4.0 release once that works. Having got this far I am considering whether
it might work for helping to evaluate whether an ISP is providing Steam HTTP
content or not (as TelstraClear do not, they have port 80 closed on their
Steam content server), but to be really useful I'd have to get some kind of
indication sent to the webservice about which way the test result fell so
even though such a thing would clearly be hugely beneficial to users I can
expect that it'd be resisted.