Add a workaround for the Avast! HTTP proxy to the port-probe utility.
This just does a minimal HTTP 1.0 HEAD request for a favicon; most sites
will return a minimal 302 response to this, while eventually the Avast!
proxy will drop the connection if the real target doesn't have an open HTTP
server port.

It's funny how some really broken sites respond, as TelstraClear's Java web
proxy would 400 the request if I requested HTTP/1.1 but succeed with an HTTP
1.1 response if I asked for a 1.0 resource. Probably the success case is the
worse one since it also includes a bunch of useless tracking cookies along
with the HEAD, but since the aim here is to be server-log-friendly then any
response that'll consistently 302 seems best.

This does push the binary size from 4096 to 4608, though. Horrors!