Integrate the new port probe utility into the end-to-end process.
This fixes a small path bug in the probe utility itself, and arranges for it
to be able to be invoked by an extension to the ISP-detection result from
the GAE webservice. Now the result can include a series of client-side port-
probe tests which can be run by the setfilter.js script, and depending on
results can apply alterations to the final detected filter.

This is primarily for South Africa where dual ISPs are evidently common, but
in principle could also be used to help with maintaining filters for the
Australia/NZ case where the ISPs do not usually document whether they are
supporting HTTP download from their customer-filtered servers. Given that
the only way to detect HTTP support is to probe for it from the client side
this potentially provides a useful way to improve filter quality.

In this build, the results of the extended probe are sent back (since this
is a beta build, and I need to be able to see that things are working as
intended in the test country); something like this really needs to be
optional if this is ever to be properly useful for detecting when ISPs add
HTTP support to their content servers.