Resolve another misfeature of the Avast! proxy in the probe assistant.
When using the probe executable to test whether port 80 is open on a target
(to help with detecting whether filtered Steam content servers can support
the new HTTP download system or not) sometimes my workaround for the Avast!
proxy wasn't working.

Turns out that sometimes the Avast! proxy returns a bogus 200 success doc
for an unreachable target and *then* closes the connection; the headers have
a refresh set to try and get the browser to retry, which will then fail in
the normal way. I detect the bogus refresh header, which no legitimate server
would ever have set (especially for a favicon, of all things).