Since NNTP probing for WebAfrica wasn't effective, try something more brutal
It transpires that the WebAfrica NNTP server is not inside WebAfrica's
special unmetered range (the "Freezone") and so while the test works it's
not the perfect way to detect dual-ISP situations since it may not be in the
static route set up by people using it.

So, I'll try something more direct; actually doing the moral equivalent of a
traceroute to the WA steam servers and trying to do a reverse DNS lookup to
find the first host on the path with a name - if that's a WA host, then the
route will be through WA's customer side and so we have a hit. If not, then
not.

While I hadn't wanted to take this route it only expands the probe code by a
few KB, and it isn't subject to interference by proxies like Avast! although
it might be more subject to interference at the packet level by firewalls.
It should also be fairly reliably fast to make a decision one way or another
whereas the NNTP-type test could be slow, so there's that.

It should be that IcmpSendEcho doesn't require elevation under UAC, since it
is used by the actual tracert.exe utility and that doesn't (whereas SOCK_RAW
sockets would require elevation).