More polish, installer tweaks and ping time comparison in the probe tool
This mainly adds a system where the probe tool which can be asked to test
the client-side network view can also compare ping times to host to help
decide between alternate rules to help detect the ideal rule set.

In addition, this makes the reporting of the rule results to the webservice
optional, since many of the rule tests aren't important to log (some, like
whether port 80 on a filtered steam service is open, need to be sent back
in order to tune the default rules).

The installer is still under 90kb, but it's getting close so one option to
help keep the install small is to minify the setfilter.js file (which saves
about 5kb). However, I'd rather deliver it in original source form as long
as possible, just so it's clearer to people what it's doing.