Adjust the rules for Adam and iiNet to reflect the latest available data.
Thanks to a heads-up from a rule submission, it's documented now that the
iiNet/Internode connection now extends to both of them having the same
free zone for Steam (which means that iiNet and Internode have the same
rule and the already-extensive Internode rule gets the 3fl servers).

See http://www.3fl.net.au/2012/03/23/internode-joins-iinets-freezone/

I've also gone through the Adam list of documented unmetered IPs carefully
to verify that the Internode server at steam.cdn.on.net is inside one of
the IP ranges that is documented as being unmetered, so I've added that
and a suitable DNS rule for getting HTTP download from those.

Since the Internode rule is so big now, I've reformatted most of the more
complex rules to at least make them slightly more comprehensible.