Refresh the IP-to-ASN mapping data and make a small adjustment to it.
The April update of the ASN mapping data is out, and since I have had a big
influx of users from iiNet following the change to their peering arrangement
with Internode and an endorsement from Whirlpool user "Frey", adding up to a
big surge in downloads that has revealed a slight quirk in the ASN mapping
data.

Lots of ISP data has small "holes", usually due to IP addresses resold by
ISPs to customers who register them as their own autonomous systems, but it
seems that the Maxmind data and the Hurricane Electric data don't quite line
up there. Holes in the Maxmind data for iiNet which don't show any ASN as
the owner at all are in use and Hurricane Electric show them as iiNet so I
am now merging adjacent lines in the source database even if there is a
small gap of unassigned IP addresses.

Right now I dont limit the size of the gap, since IPv4 usage is still at a
critical density level and any real gap would be reassigned to someone, so
most of the gaps outside the unassigned local-use ranges are not real.