Improve the diagnostic output for HTTP tracing to diagnose a problem.
I've had a report of a case where steam-limiter isn't correctly helping out
with some HTTP download in one of the more complicated Australian situations
and I haven't been able to replicate it easily, so to avoid having to go to
using Wireshark (mainly because it's an intrusive thing to ask people to
install and send me traces from) instead I'm adjusting some of the HTTP
trace system so I can get a better sense of what's going on.

So now, this uses getpeername () along with parsing out the Host: header
that is sent by the Steam client (or its embedded copy of Chromium) along
with requests so I can include that in the DebugView-capturable log. With
these two pieces of information it's easier to see what's what.

What I'm leading up to is that I can also probably match the Host: header
to add an extra layer of filtering; right now, the wonky "CS" download type
done by Steam doesn't do DNS lookup, *and* it goes through some extra URLs
like /initsession/ and /authdepot/ - so I've been using the /initsession/
URL pattern as a built-in rule as a way to knock those out.

It may well be that matching a host header and/or the host header and the
URL could be better in that respect, and having all this specific data in
the debug log will be a big help with deciding whether I want to do that or
not as I dig further into the current case.

This also adds support for building the project with VS2012 (set up for x64
hosts at the moment with the part to NSIS), which is much more convenient
when debugging on x64 hosts thanks to the improvements in x64 debugging in
VS2012. I have the projects configured to use the Windows XP targeting
support in VS2012 Update 1 so they should work just fine as beta builds,
and I'll be testing that over the next little while too.