Make the debug output for redirection show the full redirected address.
Before I printed the content of the "replace" specification when changing
the target IP underneath a connection attempt, but that spec could have a
wildcard port value which I'd print as 0. On reflection, I'd rather see the
full rewritten address instead so I print the data from "temp" which merges
the original and rewrite spec.

Also some minor gravel in copyright dates, etc.