Create rules to try and address some data leakage for australian ISPs
With the assistance of Whirlpool user networkMe it appears that the Valve-
operated servers which are used to get /serverlist/XX/YY documents are not
only fully dynamically generating those documents, they are including an
extra set of "CDN" type servers over and above the basic universal name
that has set to resolve to different servers via GeoIP looking in Valve's
custom DNS server.

This is something I had anticipated would happen, but I wasn't sure when
it would really have an effect, and what the consequences would be. As it
stands, steam.cdn.on.net and cdn.ix.asn.au are used as names returned in
the serverlist document and the new logging in the test builds shows those
names turning up in the Host field of some of the HTTP traffic as you'd
expect.

[ cdn.ix.asn.au is an alias for the server also known as steam.waia.asn.au ]

As it stands now, the majority of Australian ISPs unmeter either the Steam
servers on WAIX or the steam.cdn.on.net server, or both. So what I'm going
to do for now is add a DNS redirect for cdn.ix.asn.au to point at the on.net
server; that server doesn't seem to use any Virtual Host facility and thus
accepts cdn.ix.asn.au in the Host: field, thus a DNS redirect is enough and
I don't need to do more rewriting.

More generally, this shows I probably need to start planning out in detail
a system for replacing the /serverlist document.

This is complicated by non-blocking/overlapped I/O. There's a note in the
source from earlier investigations that the classic CDN download code uses
an OVERLAPPED structure, but I need to confirm in detail the exact API use
in the current HTTP CDN download code in the client, because if it uses the
O_NONBLOCK mode of sockets I have one set of challenges to overcome and if
it uses an OVERLAPPED structure I have a different set of challenges.

The other issue to the API-structure one is whether to truly replace the
document in situ in the HTTP stream or not. Doing it properly means fully
parsing the reponse document HTTP headers to get the Content-Length at a
minimum, as well as potentially having to deal with a Content-Encoding,
although at present the Valve servers seem to not use Content-Encoding.

If I use some form of state tracking on the SOCKET handle, then I can do
recognition on the /serverlist/ request URL to mark the SOCKET as replaced,
which has the advantage that I can use a super-simple context structure as
in general it's only possible for one such request to be outstanding at a
time.

Also if I do that, then the emulation of the response body is simpler; I
can send back the emulated response to the caller immediately without the
pass-through to read the underlying socket at all, and indeed I can force
the socket closed (because the fake response can tell the caller that the
Connection: keep-alive behaviour isn't happening and force a disconnect,
so I never need to parse the original response headers to find out how big
the original response body is to edit it out from what the caller sees).

I have to decide whether to attempt to support do that next or just push
these new rules live along with a v0.5.5 containing the new logging code.