Add a pair(!) of IPv6 addresses now used by iiNet and given to customers
Both of these IPv6 netblocks have turned up in the webserver logs for the
configuration webservice, so I'm adding them in.

I've also added an explicit shootdown rule (similar to the new one that is
going in for steam.ix.asn.au) for steam.cdn.on.net to the iPrimus and
Optus servers since Steam are also explicitly advertising that now as well
as steam.ix.asn.au, and on Optus and iPrimus the on.net servers are metered.

Still testing all these new rules so they haven't been deployed live yet,
but that should be soon.