Found a document actually describing unmetered support at Westnet
Man, ISPs really are terrible at giving out this information. Anyway, now
that I've found it I've adjusted the rule for HTTP for them to be a hybrid
of the Internode and iiNet ones since Westnet support both the WAIX and
steam.cdn.on.net servers according to their help pages.