Change the way the "no traffic" filter works, so it only blocks downloads.
This also makes the "no traffic" filter do URL blocking on the /depot/*
pattern so that no HTTP downloads work at all. The old version of the
"no traffic" filter didn't try to stop any additional CDN servers that
Value advertised, which now they have started to do. This takes care of
that.

Longer-term the intent is still to replace the /serverlist/ document, but
since it's easier to fix up the old system quickly I'm getting that done
first.