New for testing 0.5.4.5, several bugfixes and a whole new subsystem.
This now includes a new extension to the filter rules which allow an HTTP
request URL to be substituted with a replacement. For now the replacement
text is sourced from the registry, although I could use a file instead in
future easily enough (the main value of the registry is that text encoding
is better-defined there).

This involves capturing a swathe of extra APIs, some of which turned out
to be not quite what I thought; as the substituted request is not sent
over the HTTP channel there is no response body received, which I'd need to
edit out of the received stream which is potentially quite difficult. This
causes a complication with notifying the host application that the alternate
data I have in memory is ready to be read, and chasing down which API to
use for that involved a lot of trial and error, ending up with me building
a whole subsystem for dealing with WSAEventSelect () and friends which was
my first guess to eventually stumbling on select () as what the Steam code
was using instead.

The intent here is that in future I can use a rule element like:
   /serverlist/*=<generic
to swap a very generic server list in place of the one Valve actually
generate, to help avoid problems caused by Valve's new usage of dynamic
serverlist content with explicit pointers to certain content servers
which had caused some leakage.

Although I've managed to add enough new rules for now to cope with the
steam.ix.asn.au server being directly advertised, this new approach is
potentially a way to be more leakproof.

It is however, not foolproof; although the Steam servers return the
/serverlist document with an immediate cache expiry the Steam client does
appear to cache this list for a long time as a hard-coded behaviour,
which makes it tricky to do the replacement if users enable/disable the
filter without actually fully restarting Steam.

Also because this is such a complicated new system there is a question
over how I am going to configure its use; I will probably have to think
of some new mechanism in setfilter.js and the GAE webservice to deal
with the versioning issue so that 0.5.4 and older clients are never asked
to use a redirection rule. In the meantime it's here for testing purposes
and manual use, but for a full 0.5.6 release I really should have all the
ducks in a row to be able to enable it through the webservice.

The major bugfix also included in this is to work around a problem caused
by APPHELP.DLL which is a mostly-undocumented Windows component used by
the shim engine in Windows to install appcompat fixes. It's of no use to
steam-limiter, but it seems that SHLWAPI.DLL ends up triggering this to
be loaded, with dire consequences.

Normally one of the cast-iron guarantees in Windows is that even with ASLR
in place, system DLLs load at the same address in all processes of the
same bitness; this is crucial to making CreateRemoteThread () useful, as
otherwise remotely injecting code would be nearly impossible.

However, APPHELP.DLL uses IAT patching to redirect GetProcAddress to
itself; this is generally manageable since it turns out that Steam also
tends to end up with APPHELP.DLL loaded at some point, but it is very
nhelpful when Steam is first starting out because it can cause the shim
to try and call an address which isn't yet populated in a newly minted
process. And so, under some circumstances, Steam could fail if it started
with the filter running.

I've worked around this by going underneath GetProcAddress () to the inner
NTDLL.DLL function which actually implements module name lookup,
LdrGetProcedureAddress () - this lets the built shim be robust whether
the source process or target process have the APPHELP.DLL loaded or not.

There's some other gravel in the form of copyright notice updates (I'd
been lazy and not set them precisely for some files) and fixing some tab
damage in the NSIS installer script.